+++ admin/core/Driver.php
@@ -298,10 +298,21 @@ abstract class Driver
 	 */
 	public function insert(string $table, array $record)
 	{
-		return queries("INSERT INTO " . table($table) . ($record
-			? " (" . implode(", ", array_keys($record)) . ")\nVALUES (" . implode(", ", $record) . ")"
-			: " DEFAULT VALUES"
-		));
+		return queries("INSERT INTO " . table($table) . (
+			$record ?
+				" (" . implode(", ", array_keys($record)) . ")\nVALUES (" . implode(", ", $record) . ")" :
+				" DEFAULT VALUES"
+			) . $this->getInsertReturningSql($table));
+	}
+
+	/**
+	 * Returns RETURNING clause for INSERT queries (PostgreSQL specific).
+	 *
+	 * @param string $table Table name.
+	 */
+	public function getInsertReturningSql(string $table): string
+	{
+		return "";
 	}
 
 	/**